pyproject.toml
[project]
name = "segmentation"
version = "1.0.0"
description = "Point cloud segmentation prototype"
authors = [
    { name = "Murali Dhanakoti", email = "dhanakoti.murali@gmail.com" }
]
license = { text = "GPL-3.0" }
readme = "README.md"
requires-python = ">=3.13,<3.14"
dependencies = [
    "ifcopenshell (>=0.8.4.post1,<0.9.0)",
    "math3d-py (>=1.3.1,<2.0.0)",
    "vtk (>=9.5.2,<10.0.0)",
    "rich (>=14.2.0,<15.0.0)",
    "pye57 (>=0.4.19,<0.5.0)",
    "numpy (>=2.4.0,<3.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[virtualenvs]
in-project = true

[project.scripts]
segmentation = "segmentation.segment:main"
